% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blot.R
\name{blot}
\alias{blot}
\title{Matrix blotter}
\usage{
blot(X, ut = FALSE, lt = FALSE)
}
\arguments{
\item{X}{an asymmetric square matrix}

\item{ut}{logical: include the upper triangle}

\item{lt}{logical: include lower triangle}
}
\value{
a symmetric square matrix
}
\description{
Matrix blotter
}
\examples{
mat <-matrix(c(0,1,NA,1,
               NA,0,NA,NA,
               1,1,0,NA,
               NA,NA,NA,0), nrow=4, ncol=4)
mat
blot(mat)
}
\author{
Zhian N. Kamvar
}

