% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blotmat-package.r
\docType{package}
\name{blotr}
\alias{blotr}
\alias{blotr-package}
\title{blotr}
\description{
blotr
}
\details{
This R package folds matrices to make them symmetrical. There is
only one function and it's called \code{\link{blot}}.
}

